Expand the game introduction
Use the first paragraph of the introduction chapter to present a
high-level view of the game. The core themes are also touched upon, so
that a reader can keep them in mind whilepresent the game while perusing
the book.

The master file is also updated with some structural comments to better
guide future developments.